refactor: rename sketch types to match documention
